requirements.txt
streamlit>=1.30,<2
streamlit-autorefresh>=1.0.1
websocket-client>=1.6,<2
plotly>=5.18,<6
pandas>=2.0,<3
